tweak(Setup/Mysql): only unlink if file exists

See merge request tine20/tine20!3250

+++ tine20/Setup/Backend/Mysql.php
@@ -473,7 +473,9 @@ class Setup_Backend_Mysql extends Setup_Backend_Abstract
 
             exec($cmd);
         } finally {
-            unlink($mycnf);
+            if (file_exists($mycnf)) {
+                unlink($mycnf);
+            }
         }
 
         if (! $option['novalidate']) {
